d/{control,rules}: build-depend on llvm-defaults packages
Now that a liblldb-dev package exists, build-depend on the default
llvm packages instead of packages from a specific llvm version. This
avoids a new source upload every time the default llvm version changes.

To accommodate this, also change the rules file to use llvm-config to
obtain the paths to libclang.so and liblldb.so.